requirements.txt
gevent
pyyaml >= 3.11
mysql-connector >= 2.2.9
python-systemd